requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
numpy==1.18.4
pandas==1.0.3
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
